module_30_ci_linters/hw/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pytest==7.4.0
factory-boy==3.2.1
Faker==19.1.0